+++ R/shape.R
@@ -30,7 +30,7 @@ test_shape = function(shape, null_ok = FALSE, unknown_batch = NULL, len = NULL) 
   if (is.null(shape) && null_ok) {
     return(TRUE)
   }
-  ok = test_integerish(shape, min.len = 2L, all.missing = FALSE, any.missing = TRUE, len = len)
+  ok = test_integerish(shape, min.len = 1L, any.missing = TRUE, len = len)
 
   if (!ok) {
     return(FALSE)